Conda guidance
This adds instructions on how to use `conda` to manage dependencies.

It still needs fleshing out for how to use `conda` for `airflow-dags` and
`shiny` apps.
This will be done once I figure out how to make the shiny-template repo work in
a backwards compatible way. It will need to support:
 - Packrat projects (no environment.yml)
 - Packrat project using conda
 - pure conda project